Move aflow redefs to new file
